+++ dojo/filters.py
@@ -57,7 +57,6 @@ from dojo.location.status import FindingLocationStatus, ProductLocationStatus
 from dojo.models import (
     EFFORT_FOR_FIXING_CHOICES,
     ENGAGEMENT_STATUS_CHOICES,
-    IMPORT_ACTIONS,
     SEVERITY_CHOICES,
     App_Analysis,
     ChoiceQuestion,
@@ -79,7 +78,6 @@ from dojo.models import (
     Risk_Acceptance,
     Test,
     Test_Import,
-    Test_Import_Finding_Action,
     Test_Type,
     TextQuestion,
     User,
@@ -3705,47 +3703,8 @@ class UserFilter(DojoFilter):
         fields = ["is_superuser", "is_staff", "is_active", "first_name", "last_name", "username", "email"]
 
 
-# This class is used exclusively by Findings
-class TestImportFilter(DojoFilter):
-    version = CharFilter(field_name="version", lookup_expr="icontains")
-    version_exact = CharFilter(field_name="version", lookup_expr="iexact", label="Version Exact")
-    branch_tag = CharFilter(lookup_expr="icontains", label="Branch/Tag")
-    build_id = CharFilter(lookup_expr="icontains", label="Build ID")
-    commit_hash = CharFilter(lookup_expr="icontains", label="Commit hash")
-
-    findings_affected = BooleanFilter(field_name="findings_affected", lookup_expr="isnull", exclude=True, label="Findings affected")
-
-    o = OrderingFilter(
-        # tuple-mapping retains order
-        fields=(
-            ("date", "date"),
-            ("version", "version"),
-            ("branch_tag", "branch_tag"),
-            ("build_id", "build_id"),
-            ("commit_hash", "commit_hash"),
-
-        ),
-    )
-
-    class Meta:
-        model = Test_Import
-        fields = []
-
-
-# This class is used exclusively by Findings
-class TestImportFindingActionFilter(DojoFilter):
-    action = MultipleChoiceFilter(choices=IMPORT_ACTIONS)
-    o = OrderingFilter(
-        # tuple-mapping retains order
-        fields=(
-            ("action", "action"),
-        ),
-    )
-
-    class Meta:
-        model = Test_Import_Finding_Action
-        fields = []
-
+# TestImportFilter and TestImportFindingActionFilter live in dojo/test/ui/filters.py and are
+# re-exported at the bottom of this module for backward compatibility.
 
 # Used within the TestImport API
 class TestImportAPIFilter(DojoFilter):
@@ -3778,22 +3737,7 @@ class TestImportAPIFilter(DojoFilter):
 
 # LogEntryFilter and PgHistoryFilter live in dojo/auditlog/filters.py and are
 # re-exported at the bottom of this module for backward compatibility.
-
-
-class TestTypeFilter(DojoFilter):
-    name = CharFilter(lookup_expr="icontains")
-
-    o = OrderingFilter(
-        # tuple-mapping retains order
-        fields=(
-            ("name", "name"),
-        ),
-    )
-
-    class Meta:
-        model = Test_Type
-        exclude = []
-        include = ("name",)
+# TestTypeFilter lives in dojo/test/ui/filters.py and is re-exported below.
 
 
 class DevelopmentEnvironmentFilter(DojoFilter):
