+++ dojo/finding/views.py
@@ -41,8 +41,6 @@ from dojo.filters import (
     SimilarFindingFilter,
     SimilarFindingFilterWithoutObjectLookups,
     TemplateFindingFilter,
-    TestImportFilter,
-    TestImportFindingActionFilter,
 )
 from dojo.finding.deduplication import (
     _fetch_fp_candidates_for_batch,
@@ -97,6 +95,7 @@ from dojo.models import (
 from dojo.notifications.helper import create_notification
 from dojo.tags.utils import bulk_add_tags_to_instances
 from dojo.test.queries import get_authorized_tests
+from dojo.test.ui.filters import TestImportFilter, TestImportFindingActionFilter
 from dojo.tools import tool_issue_updater
 from dojo.utils import (
     FileIterWrapper,
